plant_ai_service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
python-multipart>=0.0.9
pillow>=10.0.0
opencv-python-headless>=4.8.0
